requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.55.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.1.0
idna==3.4
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.3
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.4
six==1.16.0
sqlparse==0.4.4
stripe==6.6.0
swagger-spec-validator==3.0.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
